+++ src/SecureHeadersServiceProvider.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Bepsvpt\SecureHeaders;
+
+use Illuminate\Support\ServiceProvider;
+
+class SecureHeadersServiceProvider extends ServiceProvider
+{
+    /**
+     * Bootstrap the application events.
+     *
+     * @return void
+     */
+    public function boot()
+    {
+        if ($this->app->runningInConsole()) {
+            $this->publishes([
+                __DIR__.'/../config/secure-headers.php' => config_path('secure-headers.php'),
+            ], 'config');
+        }
+    }
+
+    /**
+     * Register the service provider.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        $this->mergeConfigFrom(
+            __DIR__.'/../config/secure-headers.php', 'secure-headers'
+        );
+    }
+}