+++ tests/autoload.php
@@ -1,5 +1,0 @@
-<?php
-
-require_once __DIR__.'/helpers.php';
-
-require_once __DIR__.'/../vendor/autoload.php';